% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WCEpackage.R
\docType{package}
\name{WCE-package}
\alias{WCE-package}
\title{WCE: Weighted Cumulative Exposure Models}
\description{
A flexible method for modeling cumulative effects of time-varying exposures, weighted according to their relative proximity in time, and represented by time-dependent covariates. The current implementation estimates the weight function in the Cox proportional hazards model. The function that assigns weights to doses taken in the past is estimated using cubic regression splines.
}
\author{
\strong{Maintainer}: Marie-Pierre Sylvestre \email{marie-pierre.sylvestre@umontreal.ca} (\href{https://orcid.org/0000-0001-5803-4615}{ORCID})

Other contributors:
\itemize{
  \item Marie-Eve Beauchamp [contributor]
  \item Ryan Patrick Kyle [contributor]
  \item Michal Abrahamowicz [contributor]
}

}
\keyword{internal}
