\name{wce-package}
\alias{wce-package}
\alias{wce}
\docType{package}
\title{Weighted Cumulative Exposure Models}
\description{Fit proportional hazards models with a weighted cumulative exposure, based on a weight function estimated from cubic regression b-splines. This implementation of WCE assumes that the dataset is in the interval format with one line per unit of time.}
\details{
\tabular{ll}{
Package: \tab wce\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-01-09\cr
License: \tab GPL-2\cr
}
}
\author{
Authors: Marie-Pierre Sylvestre <marie-pierre.sylvestre@umontreal.ca>, Marie-Eve Beauchamp <marie-eve.beauchamp@clinepi.mcgill.ca>, Ryan Patrick Kyle <ryan.kyle@mail.mcgill.ca>, Michal Abrahamowicz <michal.abrahamowicz@mcgill.ca>

Maintainer: Marie-Pierre Sylvestre <marie-pierre.sylvestre@umontreal.ca>
}
\references{
Sylvestre MP, Abrahamowicz M. Flexible modeling of the cumulative effects of time-dependent exposures on the hazard. Stat Med. 2009;28(27):3437-53.
Abrahamowicz M, Beauchamp M-E, Sylvestre M-P.  Comparison of alternative models for linking drug exposure with adverse effects.  Stat Med. 2012;31(11-12):1014-1030.
}
\keyword{ package }